Connection management fixed
git-svn-id: https://xenitsupport.jira.com/svn/XENMTOA/trunk/move2alf-war@1907 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ src/main/java/eu/xenit/move2alf/core/sourcesink/AlfrescoSourceSink.java
@@ -115,103 +115,138 @@ public class AlfrescoSourceSink extends SourceSink {
 
 			File document = (File) parameterMap.get(Parameters.PARAM_FILE);
 
-			// if (!ras.doesDocExist(document.getName(), remotePath)) {
 			try {
-				ras.storeDocAndCreateParentSpaces(document, mimeType,
-						remotePath, description, namespace, contentType,
-						metadata, multiValueMetadata);
-				if (acl != null) {
-					for (String aclPath : acl.keySet()) {
-
-						String parserAclPath = aclPath;
-						// add "cm:" in front of each path component
-						if (parserAclPath.startsWith("/")) {
-							parserAclPath = parserAclPath.substring(1,
-									parserAclPath.length());
-						}
-						if (parserAclPath.endsWith("/")) {
-							parserAclPath = parserAclPath.substring(0,
-									parserAclPath.length() - 1);
-						}
-
-						String remoteACLPath = basePath + parserAclPath;
-						String[] aclComponents = remoteACLPath.split("/");
-						remoteACLPath = "";
-						for (String aclComponent : aclComponents) {
-							if ("".equals(aclComponent)) {
-								remoteACLPath += "/";
-							} else if (aclComponent.contains(":")) {
-								remoteACLPath += aclComponent + "/";
-							} else {
-								remoteACLPath += "cm:" + aclComponent + "/";
-							}
-						}
-						remoteACLPath = remoteACLPath.substring(0,
-								remoteACLPath.length() - 1);
-
-						logger.debug("ACL path: " + remoteACLPath);
-
-						ras.setAccessControlList(remoteACLPath,
-								inheritPermissions, acl.get(aclPath));
-					}
-				}
-			} catch (RepositoryException e) {
-				Throwable cause = e.getCause();
-				if (cause == null) {
-					cause = e;
-				} 
-				logger.info("Message {}", cause.getMessage());
-				Writer result = new StringWriter();
-				PrintWriter printWriter = new PrintWriter(result);
-				cause.printStackTrace(printWriter);
-				String stackTrace = result.toString();
-				logger.debug("Stacktrace {}", stackTrace);
-				if (stackTrace
-						.contains("org.alfresco.service.cmr.repository.DuplicateChildNodeNameException")) {
-					if (MODE_SKIP.equals(docExistsMode)) {
-						// ignore
-					} else if (MODE_SKIP_AND_LOG.equals(docExistsMode)) {
-						logger.warn("Document " + document.getName()
-								+ " already exists in " + remotePath);
-						throw new Move2AlfException("Document " + document.getName()
-										+ " already exists in " + remotePath);
-					} else if (MODE_OVERWRITE.equals(docExistsMode)) {
-						logger.info("Overwriting document "
-								+ document.getName() + " in " + remotePath);
-						ras.updateContentByDocNameAndPath(remotePath, document
-								.getName(), document, mimeType, false);
-						if (metadata != null) {
-							ras.updateMetaDataByDocNameAndPath(remotePath,
-									document.getName(), metadata);
-							// TODO: updating multivalue metadata not supported
-							// by
-							// RRA?
-						}
+				logger.debug("Uploading file " + document.getName());
+				uploadFile(docExistsMode, ras, basePath, remotePath, mimeType,
+						namespace, contentType, description, metadata,
+						multiValueMetadata, acl, inheritPermissions, document);
+			} catch (RuntimeException e) {
+				if ("Error writing content to repository server".equals(e
+						.getMessage())) {
+					// retry
+					logger.debug("Authentication failure? Creating new RAS");
+					destroyRepositoryAccessSession();
+					ras = createRepositoryAccessSession(configuredSourceSink);
+					try {
+						logger.debug("Retrying file " + document.getName());
+						uploadFile(docExistsMode, ras, basePath, remotePath,
+								mimeType, namespace, contentType, description,
+								metadata, multiValueMetadata, acl,
+								inheritPermissions, document);
+					} catch (RuntimeException e2) {
+						logger.error(e2.getMessage(), e2);
+						throw new Move2AlfException(e2.getMessage(), e2);
 					}
 				} else {
-					throw e;
+					logger.error(e.getMessage(), e);
+					throw new Move2AlfException(e.getMessage(), e);
 				}
 			}
 
 		} catch (RepositoryAccessException e) {
 			// we end up here if there is a communication error during a session
 			logger.error(e.getMessage(), e);
-			throw new Move2AlfException(e.getMessage());
+			throw new Move2AlfException(e.getMessage(), e);
 		} catch (RepositoryException e) {
 			// we end up here if the request could not be handled by the
 			// repository
 			logger.error(e.getMessage(), e);
-			throw new Move2AlfException(e.getMessage());
+			throw new Move2AlfException(e.getMessage(), e);
 		} catch (WebServiceException e) {
 			logger.error(e.getMessage(), e);
-			throw new Move2AlfException(e.getMessage());
+			throw new Move2AlfException(e.getMessage(), e);
 		} catch (RepositoryFatalException e) {
 			logger.error("Fatal Exception", e);
 			// TODO: stop job instead of stopping tomcat
 			// System.exit(1);
-		} catch (RuntimeException e) {
-			logger.error(e.getMessage(), e);
-			throw new Move2AlfException(e.getMessage());
+		}
+	}
+
+	private void uploadFile(final String docExistsMode, final RepositoryAccessSession ras,
+			final String basePath, final String remotePath, final String mimeType,
+			final String namespace, final String contentType, final String description,
+			final Map<String, String> metadata,
+			final Map<String, String> multiValueMetadata,
+			final Map<String, Map<String, String>> acl, final boolean inheritPermissions,
+			final File document) throws RepositoryAccessException,
+			RepositoryException {
+		try {
+			ras.storeDocAndCreateParentSpaces(document, mimeType, remotePath,
+					description, namespace, contentType, metadata,
+					multiValueMetadata);
+			if (acl != null) {
+				for (String aclPath : acl.keySet()) {
+
+					String parserAclPath = aclPath;
+					// add "cm:" in front of each path component
+					if (parserAclPath.startsWith("/")) {
+						parserAclPath = parserAclPath.substring(1,
+								parserAclPath.length());
+					}
+					if (parserAclPath.endsWith("/")) {
+						parserAclPath = parserAclPath.substring(0,
+								parserAclPath.length() - 1);
+					}
+
+					String remoteACLPath = basePath + parserAclPath;
+					String[] aclComponents = remoteACLPath.split("/");
+					remoteACLPath = "";
+					for (String aclComponent : aclComponents) {
+						if ("".equals(aclComponent)) {
+							remoteACLPath += "/";
+						} else if (aclComponent.contains(":")) {
+							remoteACLPath += aclComponent + "/";
+						} else {
+							remoteACLPath += "cm:" + aclComponent + "/";
+						}
+					}
+					remoteACLPath = remoteACLPath.substring(0, remoteACLPath
+							.length() - 1);
+
+					logger.debug("ACL path: " + remoteACLPath);
+
+					ras.setAccessControlList(remoteACLPath, inheritPermissions,
+							acl.get(aclPath));
+				}
+			}
+		} catch (RepositoryException e) {
+			Throwable cause = e.getCause();
+			if (cause == null) {
+				cause = e;
+			}
+			logger.info("Message {}", cause.getMessage());
+			Writer result = new StringWriter();
+			PrintWriter printWriter = new PrintWriter(result);
+			cause.printStackTrace(printWriter);
+			String stackTrace = result.toString();
+			//logger.debug("Stacktrace {}", stackTrace);
+			if (stackTrace
+					.contains("org.alfresco.service.cmr.repository.DuplicateChildNodeNameException")) {
+				if (MODE_SKIP.equals(docExistsMode)) {
+					// ignore
+				} else if (MODE_SKIP_AND_LOG.equals(docExistsMode)) {
+					logger.warn("Document " + document.getName()
+							+ " already exists in " + remotePath);
+					throw new Move2AlfException("Document "
+							+ document.getName() + " already exists in "
+							+ remotePath);
+				} else if (MODE_OVERWRITE.equals(docExistsMode)) {
+					logger.info("Overwriting document " + document.getName()
+							+ " in " + remotePath);
+					ras.updateContentByDocNameAndPath(remotePath, document
+							.getName(), document, mimeType, false);
+					if (metadata != null) {
+						ras.updateMetaDataByDocNameAndPath(remotePath, document
+								.getName(), metadata);
+						// TODO: updating multivalue metadata not
+						// supported
+						// by
+						// RRA?
+					}
+				}
+			} else {
+				throw e;
+			}
 		}
 	}
 
@@ -219,12 +254,28 @@ public class AlfrescoSourceSink extends SourceSink {
 	public boolean exists(ConfiguredSourceSink sinkConfig, String remotePath,
 			String name) {
 		try {
-			RepositoryAccessSession ras = createRepositoryAccessSession(sinkConfig);
-			return ras.doesDocExist(name, remotePath);
-		} catch (RepositoryAccessException e) {
-			throw new Move2AlfException(e.getMessage());
+			try {
+				RepositoryAccessSession ras = createRepositoryAccessSession(sinkConfig);
+				return ras.doesDocExist(name, remotePath);
+			} catch (RepositoryAccessException e) {
+				if (!(e.getMessage() == null) && (e.getMessage().indexOf("security processing failed") != -1)) { 
+					// retry
+					destroyRepositoryAccessSession();
+					RepositoryAccessSession ras = createRepositoryAccessSession(sinkConfig);
+					try {
+						return ras.doesDocExist(name, remotePath);
+					} catch (RepositoryAccessException e2) {
+						logger.error(e2.getMessage(), e2);
+						throw new Move2AlfException(e2.getMessage(), e2);
+					}
+				} else {
+					logger.error(e.getMessage(), e);
+					throw new Move2AlfException(e.getMessage(), e);
+				}
+			}
 		} catch (RuntimeException e) {
-			throw new Move2AlfException(e.getMessage());
+			logger.error(e.getMessage(), e);
+			throw new Move2AlfException(e.getMessage(), e);
 		}
 	}
 
@@ -232,20 +283,37 @@ public class AlfrescoSourceSink extends SourceSink {
 	public void delete(ConfiguredSourceSink sinkConfig, String remotePath,
 			String name) {
 		try {
-			RepositoryAccessSession ras = createRepositoryAccessSession(sinkConfig);
-			ras.deleteByDocNameAndSpace(remotePath, name);
-		} catch (RepositoryAccessException e) {
-			throw new Move2AlfException(e.getMessage());
+			try {
+				RepositoryAccessSession ras = createRepositoryAccessSession(sinkConfig);
+				ras.deleteByDocNameAndSpace(remotePath, name);
+			} catch (RepositoryAccessException e) {
+				if (!(e.getMessage() == null) && (e.getMessage().indexOf("security processing failed") != -1)) { 
+					// retry
+					destroyRepositoryAccessSession();
+					RepositoryAccessSession ras = createRepositoryAccessSession(sinkConfig);
+					try {
+						ras.deleteByDocNameAndSpace(remotePath, name);
+					} catch (RepositoryAccessException e2) {
+						logger.error(e.getMessage(), e);
+						throw new Move2AlfException(e.getMessage(), e);
+					}
+				} else {
+					logger.error(e.getMessage(), e);
+					throw new Move2AlfException(e.getMessage(), e);
+				}
+			}
 		} catch (RepositoryException e) {
-			throw new Move2AlfException(e.getMessage());
+			logger.error(e.getMessage(), e);
+			throw new Move2AlfException(e.getMessage(), e);
 		} catch (RuntimeException e) {
-			throw new Move2AlfException(e.getMessage());
+			logger.error(e.getMessage(), e);
+			throw new Move2AlfException(e.getMessage(), e);
 		}
 	}
 
 	private RepositoryAccessSession createRepositoryAccessSession(
 			ConfiguredSourceSink sinkConfig) {
-		//RepositoryAccessSession ras;
+		// RepositoryAccessSession ras;
 		RepositoryAccessSession ras = AlfrescoSourceSink.ras.get();
 		if (ras == null) {
 			logger.debug("Creating new RepositoryAccessSession for thread "
@@ -273,7 +341,12 @@ public class AlfrescoSourceSink extends SourceSink {
 					+ Thread.currentThread());
 		}
 		return AlfrescoSourceSink.ras.get();
-		//return ras;
+		// return ras;
+	}
+
+	private void destroyRepositoryAccessSession() {
+		AlfrescoSourceSink.ras.get().closeSession();
+		AlfrescoSourceSink.ras.remove();
 	}
 
 	private String getParameterWithDefault(Map<String, Object> parameterMap,
